Backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-socketio
python-dotenv
eventlet
werkzeug
gunicorn
psycopg2-binary
matplotlib
pandas
reportlab
torch
transformers
pillow
numpy
